# add cases for unevaluated keys to MAP hooks