Some of the `MAP` hooks were missing cases for an unevaluated key argument (they were assuming that the argument would already be fully evaluated when the hook is called).
In these cases, the hooks should typically return `Nothing` instead of returning a result.